requirements.txt
streamlit=1.33.0
pytube=15.0.0
google-generativeai=0.5.2
deepgram-sdk=3.2.7
python-dotenv=1.0.1
python-docx=1.1.0